+++ lib/pats/categories/signal_quality_grades/psg_e1_signal_quality_grade.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+require 'pats/categories/default'
+
+module Pats
+  module Categories
+    module SignalQualityGrades
+      # Defines PSG E1 Signal Quality Grade Mean
+      class PSGE1SignalQualityGrade < Default
+        def label
+          'E1 Signal Quality Grade'
+        end
+
+        def variable_name
+          'psg_e1_signal_quality_grade'
+        end
+
+        def subquery
+          "#{database_value} IN (1, 2, 3, 4, 5)"
+        end
+      end
+    end
+  end
+end